Added toggle button and functionality to Performance Visx. Placed BarGraph into a separate component.
